chore: unnecessary code removed & code refactoring
