+++ src/components/charts/charts.types.tsx
@@ -19,6 +19,34 @@ import { RouteComponentProps } from 'react-router-dom'
 
 import { OptionType, ResponseType } from '@devtron-labs/devtron-fe-common-lib'
 
+export interface ChartGroupCardProps {
+    chartGroup: ChartGroup
+}
+
+export interface AllChartSelectProps {
+    chart: Chart
+    isListView: boolean
+    dataTestId: string
+    selectedCount: number
+    onClick?: (chartId: number) => void
+    addChart?: (chartId: number) => void
+    subtractChart?: (chartId: number) => void
+}
+
+export interface Stepper extends AllChartSelectProps {
+    addChart: (chartId: number) => void
+    subtractChart: (chartId: number) => void
+    selectChart?: never
+}
+
+export interface Checkbox extends AllChartSelectProps {
+    addChart?: never
+    subtractChart?: never
+    selectChart: (chartId: number) => void
+}
+
+export type ChartSelectProps = Stepper | Checkbox
+
 export interface ChartValuesType {
     kind: 'DEFAULT' | 'TEMPLATE' | 'DEPLOYED' | 'EXISTING' | null
     chartVersion?: string
