+++ src/components/charts/util/ChartGroupCard.tsx
@@ -16,10 +16,9 @@
 
 import { Component } from 'react'
 import { Link } from 'react-router-dom'
-import { ChartGroupEntry } from '../charts.types'
+import { ChartGroupCardProps, ChartGroupEntry } from '../charts.types'
 import { getChartGroupURL } from '../charts.helper'
 import placeHolder from '../../../assets/icons/ic-plc-chart.svg'
-import { ChartGroupCardProps } from './types'
 import { handleAnalyticsEvent } from '@devtron-labs/devtron-fe-common-lib'
 
 export default class ChartGroupCard extends Component<ChartGroupCardProps> {
@@ -39,7 +38,6 @@ export default class ChartGroupCard extends Component<ChartGroupCardProps> {
         const { chartGroup } = this.props
         const chartGroupEntries = this.getChartGroupSubgroup()
         const GROUP_EDIT_LINK = getChartGroupURL(chartGroup.id)
-        const classes = 'chart-grid-item h-200 white-card dc__no-decor'
 
         const handleClick = () => {
             handleAnalyticsEvent({ category: 'Chart Store', action: 'CS_CHART_GROUP_CARD_CLICKED' })
